+++ src/net/frakbot/FWeather/FWeatherWidgetProvider.java
@@ -24,19 +24,46 @@ import android.content.Intent;
 import android.content.pm.ApplicationInfo;
 import android.os.Build;
 import android.os.Bundle;
-import android.util.Log;
+import net.frakbot.FWeather.global.Const;
 import net.frakbot.FWeather.updater.UpdaterService;
+import net.frakbot.FWeather.util.WidgetHelper;
+import net.frakbot.FWeather.util.FLog;
 
 public class FWeatherWidgetProvider extends AppWidgetProvider {
 
+    private static final String TAG = FWeatherWidgetProvider.class.getSimpleName();
+
     public void onUpdate(Context context, AppWidgetManager appWidgetManager, int[] appWidgetIds) {
         final ApplicationInfo info = context.getApplicationInfo();
-        final String tag = info != null ? info.name : "FWeather";
-        Log.i(tag, "Update started");
+        final String tag = info != null ? info.name : TAG;
+        FLog.i(context, tag, "Update started");
 
+        startUpdate(context, appWidgetIds);
+    }
+
+    @Override
+    public void onReceive(Context context, Intent intent) {
+        super.onReceive(context, intent);
+
+        if (!intent.getAction().equals(Const.Intents.ACTION_UPDATE_FRAKKING_WIDGET))
+            return;
+
+        final ApplicationInfo info = context.getApplicationInfo();
+        final String tag = info != null ? info.name : TAG;
+        FLog.i(context, tag, "Update action received!");
+
+        startUpdate(context, WidgetHelper.getWidgetIds(context));
+    }
+
+    /**
+     * Immediately requests an update by starting the service.
+     * @param context       The Context
+     * @param appWidgetIds  The App Widget IDs to update
+     */
+    private void startUpdate(Context context, int[] appWidgetIds) {
         final Intent updaterIntent =
-            new Intent(context, UpdaterService.class)
-                .putExtra(UpdaterService.EXTRA_WIDGET_IDS, appWidgetIds);
+                new Intent(context, UpdaterService.class)
+                        .putExtra(UpdaterService.EXTRA_WIDGET_IDS, appWidgetIds);
 
         context.startService(updaterIntent);
     }
@@ -50,10 +77,10 @@ public class FWeatherWidgetProvider extends AppWidgetProvider {
             final ApplicationInfo info = context.getApplicationInfo();
             final String tag = info != null ? info.name : "FWeather";
 
-            Log.i(tag, "Widget options changed, updating it");
+            FLog.i(context, tag, "Widget options changed, updating it");
             final Intent updaterIntent =
-                new Intent(context, UpdaterService.class)
-                    .putExtra(UpdaterService.EXTRA_WIDGET_IDS, new int[] {appWidgetId});
+                    new Intent(context, UpdaterService.class)
+                            .putExtra(UpdaterService.EXTRA_WIDGET_IDS, new int[] {appWidgetId});
 
             context.startService(updaterIntent);
         }