+++ src/net/frakbot/FWeather/updater/UpdaterService.java
@@ -26,7 +26,6 @@ import android.content.SharedPreferences;
 import android.location.Address;
 import android.location.Geocoder;
 import android.location.Location;
-import android.location.LocationManager;
 import android.net.ConnectivityManager;
 import android.net.NetworkInfo;
 import android.os.Build;
@@ -34,21 +33,20 @@ import android.os.Bundle;
 import android.os.Handler;
 import android.preference.PreferenceManager;
 import android.text.TextUtils;
-import android.util.Log;
 import android.view.View;
 import android.widget.RemoteViews;
 import android.widget.Toast;
-import com.google.android.gms.location.LocationClient;
 import net.frakbot.FWeather.R;
 import net.frakbot.FWeather.activity.SettingsActivity;
 import net.frakbot.FWeather.updater.weather.JSONWeatherParser;
 import net.frakbot.FWeather.updater.weather.WeatherHttpClient;
 import net.frakbot.FWeather.updater.weather.model.Weather;
-import net.frakbot.FWeather.util.LocationHelper;
-import net.frakbot.FWeather.util.TrackerHelper;
+import net.frakbot.FWeather.util.*;
 import org.json.JSONException;
 
 import java.io.IOException;
+import java.io.UnsupportedEncodingException;
+import java.net.URLEncoder;
 import java.util.List;
 
 /**
@@ -59,7 +57,7 @@ import java.util.List;
 public class UpdaterService extends IntentService {
 
     public static final String TAG = UpdaterService.class.getSimpleName();
-    private WidgetUiHelper mWidgetUiHelper;
+    private WidgetHelper mWidgetHelper;
 
     public static final String EXTRA_USER_FORCE_UPDATE = "the_motherfocker_wants_us_to_do_stuff";
     public static final String EXTRA_SILENT_FORCE_UPDATE = "a_ninja_is_making_me_do_it";
@@ -73,10 +71,10 @@ public class UpdaterService extends IntentService {
     @Override
     public void onCreate() {
         super.onCreate();
-        Log.d(TAG, "onCreate");
+        FLog.d(this, TAG, "onCreate");
 
-        Log.i(TAG, "Initializing the UpdaterService");
-        mWidgetUiHelper = new WidgetUiHelper(this);
+        FLog.i(this, TAG, "Initializing the UpdaterService");
+        mWidgetHelper = new WidgetHelper(this);
         mHandler = new Handler();
 
         // Initialize the amazing LocationHelper
@@ -86,28 +84,32 @@ public class UpdaterService extends IntentService {
 
     @Override
     protected void onHandleIntent(Intent intent) {
-        Log.d(TAG, "onHandleIntent");
+        FLog.d(this, TAG, "onHandleIntent");
+
+        // First thing, recheck the log filtering levels
+        FLog.recheckLogLevels();
+
         int[] appWidgetIds = intent.getIntArrayExtra(EXTRA_WIDGET_IDS);
         if (appWidgetIds == null || appWidgetIds.length == 0) {
-            Log.d(TAG, "Intent with no widgets ID received, ignoring\n\t> " + intent);
+            FLog.d(this, TAG, "Intent with no widgets ID received, ignoring\n\t> " + intent);
             return;
         }
 
         if (intent.getBooleanExtra(EXTRA_USER_FORCE_UPDATE, false)) {
-            Log.i(TAG, "User has requested a forced update");
+            FLog.i(this, TAG, "User has requested a forced update");
             // TODO: custom Toast layout? Would be nice.
             mHandler.post(new Runnable() {
                 @Override
                 public void run() {
                     // We need this because the IntentService thread is too fast and dies too soon,
                     // resulting in the toast being on screen for an unpercievable time
-                    WidgetUiHelper.makeToast(UpdaterService.this, R.string.toast_force_update, Toast.LENGTH_LONG)
+                    WidgetHelper.makeToast(UpdaterService.this, R.string.toast_force_update, Toast.LENGTH_LONG)
                                   .show();
                 }
             });
         }
 
-        Log.i(TAG, "Starting widgets update");
+        FLog.i(this, TAG, "Starting widgets update");
         AppWidgetManager appWidgetManager = AppWidgetManager.getInstance(this);
         assert appWidgetManager != null;
 
@@ -117,13 +119,13 @@ public class UpdaterService extends IntentService {
             weather = getWeather();
         } catch (LocationHelper.LocationNotReadyYetException justWaitException) {
             // If the location is not ready yet, leave the View unchanged
-            Log.d(TAG, "The LocationHelper is not reayd yet, the updater will be called again soon.");
+            FLog.d(this, TAG, "The LocationHelper is not ready yet, the updater will be called again soon.");
             return;
         }
 
         // Perform this loop procedure for each App Widget that belongs to this provider
         for (int appWidgetId : appWidgetIds) {
-            Log.i(TAG, "Updating the widget views for widget #" + appWidgetId);
+            FLog.i(this, TAG, "Updating the widget views for widget #" + appWidgetId);
 
             // Get the widget layout and update it
             RemoteViews views = new RemoteViews(getPackageName(),
@@ -134,7 +136,10 @@ public class UpdaterService extends IntentService {
             appWidgetManager.updateAppWidget(appWidgetId, views);
         }
 
-        Log.i(TAG, "All widgets updated successfully");
+        // Reschedule the alarm
+        AlarmHelper.rescheduleAlarm(this);
+
+        FLog.i(this, TAG, "All widgets updated successfully");
     }
 
     @TargetApi(Build.VERSION_CODES.JELLY_BEAN)
@@ -181,12 +186,12 @@ public class UpdaterService extends IntentService {
         }
 
         // Show/hide elements, and update them only if needed
-        views.setTextViewText(R.id.txt_weather, mWidgetUiHelper.getWeatherString(weather, darkMode));
+        views.setTextViewText(R.id.txt_weather, mWidgetHelper.getWeatherString(weather, darkMode));
         views.setTextColor(R.id.txt_weather, textColor);
 
         if (prefs.getBoolean(getString(R.string.pref_key_ui_toggle_temperature_info), true)) {
             views.setViewVisibility(R.id.txt_temp, View.VISIBLE);
-            views.setTextViewText(R.id.txt_temp, mWidgetUiHelper.getTempString(weather, darkMode));
+            views.setTextViewText(R.id.txt_temp, mWidgetHelper.getTempString(weather, darkMode));
             views.setTextColor(R.id.txt_temp, textColor);
         }
         else {
@@ -195,10 +200,10 @@ public class UpdaterService extends IntentService {
 
         if (prefs.getBoolean(getString(R.string.pref_key_ui_toggle_weather_icon), true)) {
             views.setViewVisibility(R.id.img_weathericon, View.VISIBLE);
-            views.setImageViewResource(R.id.img_weathericon, mWidgetUiHelper.getWeatherImageId(weather, darkMode));
+            views.setImageViewResource(R.id.img_weathericon, mWidgetHelper.getWeatherImageId(weather, darkMode));
         }
         else {
-            views.setViewVisibility(R.id.img_weathericon, View.GONE);
+            views.setViewVisibility(R.id.img_weathericon, View.INVISIBLE);
         }
 
         if (prefs.getBoolean(getString(R.string.pref_key_ui_toggle_buttons), true)) {
@@ -230,18 +235,18 @@ public class UpdaterService extends IntentService {
      */
     private Weather getWeather() throws LocationHelper.LocationNotReadyYetException {
         if (!checkNetwork()) {
-            Log.e(TAG, "Can't update weather, no network connectivity available");
+            FLog.e(this, TAG, "Can't update weather, no network connectivity available");
             return null;
         }
 
-        Log.i(TAG, "Starting weather update");
+        FLog.i(this, TAG, "Starting weather update");
 
         // Get the current location
         final Location location = getLocation();
 
         if (location == null) {
             TrackerHelper.sendException(this, "No location found", false);
-            Log.e(TAG, "No location available, can't update");
+            FLog.e(this, TAG, "No location available, can't update");
             return null;
         }
 
@@ -252,15 +257,14 @@ public class UpdaterService extends IntentService {
         String json;
 
         if (!TextUtils.isEmpty(cityName)) {
-            json = ((new WeatherHttpClient()).getCityWeatherJsonData(cityName));
-        }
-        else {
+            json = ((new WeatherHttpClient(this)).getCityWeatherJsonData(cityName));
+        } else {
             // No city name available. Use latlon values instead
-            json = ((new WeatherHttpClient()).getLocationWeatherJsonData(location));
+            json = ((new WeatherHttpClient(this)).getLocationWeatherJsonData(location));
         }
 
         if (TextUtils.isEmpty(json)) {
-            Log.e(TAG, "No weather available, can't update");
+            FLog.e(this, TAG, "No weather available, can't update");
             TrackerHelper.sendException(this, "No weather data", false);
             return null;
         }
@@ -269,13 +273,13 @@ public class UpdaterService extends IntentService {
             weather = JSONWeatherParser.getWeather(json);
         }
         catch (JSONException e) {
-            Log.e(TAG, "Weather data is not valid, can't update");
+            FLog.e(this, TAG, "Weather data is not valid, can't update");
             TrackerHelper.sendException(this, "Invalid weather JSON", false);
             return null;
         }
 
-        Log.i(TAG, "Weather update done");
-        Log.v(TAG, "Got weather:\n\t> " + weather);
+        FLog.i(this, TAG, "Weather update done");
+        FLog.v(this, TAG, "Got weather:\n\t> " + weather);
         return weather;
     }
 
@@ -285,7 +289,7 @@ public class UpdaterService extends IntentService {
      * @return Returns the current location
      */
     private Location getLocation() throws LocationHelper.LocationNotReadyYetException {
-        final Intent intent = WidgetUiHelper.getUpdaterIntent(this, false, false);
+        final Intent intent = WidgetHelper.getUpdaterIntent(this, false, false);
         final PendingIntent pendingIntent = PendingIntent.getService(this, 42, intent, 0);
         return LocationHelper.getLastKnownSurroundings(pendingIntent);
     }
@@ -320,7 +324,18 @@ public class UpdaterService extends IntentService {
                 }
             }
         }
-        return cityName;
+
+        String encodedCityName = null;
+
+        try {
+            encodedCityName = URLEncoder.encode(cityName, "UTF-8");
+        } catch (UnsupportedEncodingException e) {
+            FLog.d(this, TAG, "Could not encode city name, assume no city available.");
+        } catch (NullPointerException enp) {
+            FLog.d(this, TAG, "Could not encode city name, assume no city available.");
+        }
+
+        return encodedCityName;
     }
 
     /**