+++ src/net/frakbot/FWeather/updater/weather/WeatherHttpClient.java
@@ -11,9 +11,10 @@
  */
 package net.frakbot.FWeather.updater.weather;
 
+import android.content.Context;
 import android.location.Location;
-import android.util.Log;
 import net.frakbot.FWeather.FWeatherApplication;
+import net.frakbot.FWeather.util.FLog;
 
 import java.io.*;
 import java.net.HttpURLConnection;
@@ -45,6 +46,12 @@ public class WeatherHttpClient {
                                                   FWeatherApplication.getApiKey();
     private final static String IMG_URL = "http://openweathermap.org/img/w/";
 
+    private Context mContext;
+
+    public WeatherHttpClient(Context context) {
+        mContext = context;
+    }
+
     /**
      * Gets the JSON weather data for the specified city.
      *
@@ -58,7 +65,7 @@ public class WeatherHttpClient {
         InputStream is = null;
 
         try {
-            Log.i("FWeatherHttp", "City weather update started: " + cityName);
+            FLog.i(mContext, "FWeatherHttp", "City weather update started: " + cityName);
             con = openConnection(String.format(CITY_BASE_URL, cityName));
 
             // Let's read the response
@@ -73,12 +80,12 @@ public class WeatherHttpClient {
             is.close();
             con.disconnect();
 
-            Log.d("FWeatherHttp", "City weather update done. Length: " + buffer.length());
+            FLog.d(mContext, "FWeatherHttp", "City weather update done. Length: " + buffer.length());
 
             return buffer.toString();
         }
         catch (Throwable t) {
-            Log.e("FWeatherHttp", "Error while fetching the weather", t);
+            FLog.e(mContext, "FWeatherHttp", "Error while fetching the weather", t);
         }
         finally {
             cleanup(con, is);
@@ -100,7 +107,7 @@ public class WeatherHttpClient {
         InputStream is = null;
 
         try {
-            Log.i("FWeatherHttp", "Loaction weather update started: " + location.toString());
+            FLog.i(mContext, "FWeatherHttp", "Loaction weather update started: " + location.toString());
             // We specify the locale to avoid passing latlong data with the wrong decimal separators
             con = openConnection(String.format(Locale.US, LATLON_BASE_URL, location.getLatitude(),
                                                location.getLongitude()));
@@ -117,12 +124,12 @@ public class WeatherHttpClient {
             is.close();
             con.disconnect();
 
-            Log.d("FWeatherHttp", "Location weather update done. Length: " + buffer.length());
+            FLog.d(mContext, "FWeatherHttp", "Location weather update done. Length: " + buffer.length());
 
             return buffer.toString();
         }
         catch (Throwable t) {
-            Log.e("FWeatherHttp", "Error while fetching the weather", t);
+            FLog.e(mContext, "FWeatherHttp", "Error while fetching the weather", t);
         }
         finally {
             cleanup(con, is);
@@ -180,7 +187,7 @@ public class WeatherHttpClient {
         InputStream is = null;
 
         try {
-            Log.i("FWeatherHttp", "Image DL started. Code: " + code);
+            FLog.i(mContext, "FWeatherHttp", "Image DL started. Code: " + code);
 
             con = (HttpURLConnection) (new URL(IMG_URL + code)).openConnection();
             con.setRequestMethod("GET");
@@ -197,11 +204,11 @@ public class WeatherHttpClient {
                 baos.write(buffer);
             }
 
-            Log.i("FWeatherHttp", "Image DL done. Size: " + baos.size());
+            FLog.i(mContext, "FWeatherHttp", "Image DL done. Size: " + baos.size());
             return baos.toByteArray();
         }
         catch (Throwable t) {
-            Log.e("FWeatherHttp", "Error while fetching the weather image", t);
+            FLog.e(mContext, "FWeatherHttp", "Error while fetching the weather image", t);
         }
         finally {
             cleanup(con, is);